pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medialyze"
version = "0.1.0"
description = "Self-hosted media library analyzer for technical video metadata."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Frederik Emmer" }
]
dependencies = [
  "alembic>=1.14.1,<2",
  "apscheduler>=3.10.4,<4",
  "fastapi>=0.116.1,<1",
  "pydantic-settings>=2.8.1,<3",
  "sqlalchemy>=2.0.39,<3",
  "uvicorn[standard]>=0.34.0,<1",
  "watchdog>=6.0.0,<7",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1",
  "pytest>=8.3.5,<9",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"
